+++ mobile_wallet/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.10.0
+  - New functions `create_configure_baker_transaction` and `create_configure_delegation_transaction`
+    have been added to support the new transaction types `configure baker` and `configure delegation`
+    that are introduced in Protocol Version 4.
+  - A new function `generate_baker_keys` has been added.
+
 ## 0.9.0
    - The functions `create_transfer` and `create_encrypted_transfer` have been extended to support
      the new transfer types, i.e. transfer with memo and encrypted transfer with memo, respectively.